@@ -8,21 +8,23 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "build-prod": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --prod"
+    "build-prod": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --prod",
+    "api-gen": "ng-openapi-gen --input ./src/openApi/openApi.json  --output ./src/app/services"
   },
   "private": true,
   "dependencies": {
     "@agm/core": "^1.1.0",
-    "@angular/animations": "^13.0.2",
-    "@angular/common": "~13.0.2",
-    "@angular/compiler": "~13.0.2",
-    "@angular/core": "~13.0.2",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/fire": "^6.1.5",
-    "@angular/forms": "~13.0.2",
-    "@angular/localize": "^13.0.2",
-    "@angular/platform-browser": "~13.0.2",
-    "@angular/platform-browser-dynamic": "~13.0.2",
-    "@angular/router": "~13.0.2",
+    "@angular/forms": "^14.3.0",
+    "@angular/localize": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@auth0/angular-jwt": "^5.2.0",
     "@ckeditor/ckeditor5-angular": "^2.0.1",
     "@ckeditor/ckeditor5-build-classic": "^26.0.0",
     "@fullcalendar/angular": "^5.8.0",
@@ -37,19 +39,24 @@
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@nicky-lenaers/ngx-scroll-to": "^9.0.0",
+    "@stomp/stompjs": "^7.0.0",
     "@types/chartist": "^0.11.0",
     "angular-archwizard": "^7.0.0",
+    "angular-code-input": "^1.0.0",
     "apexcharts": "^3.27.1",
     "bootstrap": "^5.1.3",
     "bs-custom-file-input": "^1.3.4",
     "chart.js": "^2.9.4",
     "chartist": "^0.11.4",
+    "dropzone": "^5.9.3",
     "echarts": "^5.1.2",
     "firebase": "^8.6.8",
+    "jwt-decode": "^4.0.0",
     "metismenujs": "^1.2.1",
     "ng-apexcharts": "^1.5.12",
     "ng-chartist": "^4.1.0",
     "ng-click-outside": "^8.0.0",
+    "ng-openapi-gen": "^0.51.0",
     "ng-otp-input": "^1.7.1",
     "ng2-charts": "^2.4.2",
     "ng2-completer": "^9.0.1",
@@ -67,21 +74,23 @@
     "ngx-lightbox": "^2.4.0",
     "ngx-mask": "^12.0.0",
     "ngx-owl-carousel-o": "^6.0.0",
+    "ngx-socket-io": "^4.7.0",
     "ngx-ui-switch": "^10.0.2",
     "ngx-youtube-player": "^9.1.0",
     "resize-observer-polyfill": "^1.5.1",
-    "rxjs": "~6.6.6",
+    "rxjs": "^6.0.0",
     "sass-loader": "^11.1.1",
     "simplebar-angular": "^2.3.4",
+    "sockjs-client": "^1.6.1",
     "sweetalert2": "^11.0.18",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.0.3",
-    "@angular/cli": "^13.0.3",
-    "@angular/compiler-cli": "~13.0.2",
-    "@angular/language-service": "~13.0.2",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@juggle/resize-observer": "^3.3.1",
     "@types/echarts": "^4.9.8",
     "@types/jasmine": "~3.6.0",
@@ -97,6 +106,6 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   }
 }
