@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { AuthenticationService } from '../../../core/services/auth.service';
 import { AuthfakeauthenticationService } from '../../../core/services/authfake.service';
-
 import { ActivatedRoute, Router } from '@angular/router';
 import { first } from 'rxjs/operators';
-
 import { environment } from '../../../../environments/environment';
+import { AuthenticationRequest } from 'src/app/services/models';
+import { AuthenticationControllerService } from 'src/app/services/services';
+import { TokenService } from 'src/app/services/token/token.service';
 
 @Component({
   selector: 'app-login',
@@ -20,62 +20,80 @@ import { environment } from '../../../../environments/environment';
  */
 export class LoginComponent implements OnInit {
 
-  loginForm: FormGroup;
+  loginForm: UntypedFormGroup;
   submitted = false;
   error = '';
-  returnUrl: string;
-
+  authRequest: AuthenticationRequest= {email: '', password: ''};
+  errorMsg: Array<string> = [];
   // set the currenr year
   year: number = new Date().getFullYear();
 
   // tslint:disable-next-line: max-line-length
-  constructor(private formBuilder: FormBuilder, private route: ActivatedRoute, private router: Router, private authenticationService: AuthenticationService,
-    private authFackservice: AuthfakeauthenticationService) { }
+  constructor(
+    private formBuilder: UntypedFormBuilder, 
+    private route: ActivatedRoute, 
+    private router: Router,
+    private authService:AuthenticationControllerService,
+    private tokenservice:TokenService,) { }
 
-  ngOnInit() {
-    this.loginForm = this.formBuilder.group({
-      email: ['admin@themesbrand.com', [Validators.required, Validators.email]],
-      password: ['123456', [Validators.required]],
-    });
-
-    // reset login status
-    // this.authenticationService.logout();
-    // get return url from route parameters or default to '/'
-    // tslint:disable-next-line: no-string-literal
-    this.returnUrl = this.route.snapshot.queryParams['returnUrl'] || '/';
-  }
-
-  // convenience getter for easy access to form fields
-  get f() { return this.loginForm.controls; }
-
+    ngOnInit(): void {
+      this.loginForm = this.formBuilder.group({
+        email: ['', [Validators.required, Validators.email]],
+        password: ['', Validators.required]
+      });
+    }
+  
+    get f() {
+      return this.loginForm.controls;
+    }
   /**
    * Form submit
    */
   onSubmit() {
     this.submitted = true;
-
-    // stop here if form is invalid
     if (this.loginForm.invalid) {
       return;
     } else {
-      if (environment.defaultauth === 'firebase') {
-        this.authenticationService.login(this.f.email.value, this.f.password.value).then((res: any) => {
-          this.router.navigate(['/dashboard']);
-        })
-          .catch(error => {
-            this.error = error ? error : '';
-          });
-      } else {
-        this.authFackservice.login(this.f.email.value, this.f.password.value)
-          .pipe(first())
-          .subscribe(
-            data => {
-              this.router.navigate(['/dashboard']);
-            },
-            error => {
-              this.error = error ? error : '';
-            });
-      }
+      this.authRequest.email = this.f.email.value;
+    this.authRequest.password = this.f.password.value;
+        this.authService.authenticate$Response({
+          body: this.authRequest
+        }).subscribe({
+          next: (response) => {
+            if (response.body instanceof Blob) {
+              response.body.text().then(text => {
+                try {
+                  const jsonResponse = JSON.parse(text);
+                  console.log('Full Response:', jsonResponse);
+                  console.log('Token:', jsonResponse.token); // Accès au token
+                  this.tokenservice.token = jsonResponse.token;
+                  this.router.navigate(['']);
+                } catch (err) {
+                  console.error('Erreur lors de la conversion du Blob en JSON:', err);
+                }
+              }).catch(err => {
+                console.error('Erreur lors de la lecture du Blob:', err);
+              });
+            } else {
+              console.error('La réponse n\'est pas un Blob:', response.body);
+            }
+          },
+          error: (err) => {
+          
+            console.error('Error:', err);
+
+            if (err.error && err.error.validationErrors) {
+              this.error = 'Erreur de validation : ' + err.error.validationErrors.join(', ');
+            } else if (err.error && err.error.error) {
+              this.error = err.error.error; // Message d'erreur spécifique du backend
+            } else if (err.status === 404) {
+              this.error = 'Utilisateur non trouvé. Veuillez vérifier vos informations de connexion.';
+            } else {
+              this.error = 'Erreur inconnue. Veuillez réessayer plus tard.';
+            }
+          }
+        });
+      }}
+
     }
-  }
-}
+
