@@ -1,12 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 
 import { OwlOptions } from 'ngx-owl-carousel-o';
-import { AuthenticationService } from '../../../core/services/auth.service';
+import { AuthenticationControllerService } from 'src/app/services/services';
 import { environment } from '../../../../environments/environment';
 import { first } from 'rxjs/operators';
 import { UserProfileService } from '../../../core/services/user.service';
+import { RegistrationRequest } from 'src/app/services/models';
 
 @Component({
   selector: 'app-register2',
@@ -14,13 +15,17 @@ import { UserProfileService } from '../../../core/services/user.service';
   styleUrls: ['./register2.component.scss']
 })
 export class Register2Component implements OnInit {
-
-  signupForm: FormGroup;
+  registerRequest:RegistrationRequest = {email: '', firstname: '', lastname: '', password: ''};
+  signupForm: UntypedFormGroup;
   submitted = false;
   error = '';
   successmsg = false;
+ 
 
-  constructor(private formBuilder: FormBuilder, private route: ActivatedRoute, private router: Router, private authenticationService: AuthenticationService,
+  constructor(private formBuilder: UntypedFormBuilder, 
+    private route: ActivatedRoute, 
+    private router: Router,
+    private authService:AuthenticationControllerService,
     private userService: UserProfileService) { }
   // set the currenr year
   year: number = new Date().getFullYear();
@@ -30,6 +35,7 @@ export class Register2Component implements OnInit {
 
     this.signupForm = this.formBuilder.group({
       username: ['', Validators.required],
+      lastname: ['', Validators.required],
       email: ['', [Validators.required, Validators.email]],
       password: ['', Validators.required],
     });
@@ -61,30 +67,20 @@ export class Register2Component implements OnInit {
     if (this.signupForm.invalid) {
       return;
     } else {
-      if (environment.defaultauth === 'firebase') {
-        this.authenticationService.register(this.f.email.value, this.f.password.value).then((res: any) => {
-          this.successmsg = true;
-          if (this.successmsg) {
-            this.router.navigate(['/dashboard']);
+      // console.log(this.registerRequest.email)
+      // this.errorMsg = [];
+      this.authService.register({
+        body: this.registerRequest
+      })
+        .subscribe({
+          next: () => {
+            this.router.navigate(['']);
+          },
+          error: (err) => {
+            this.error = err.error.validationErrors;
           }
-        })
-          .catch(error => {
-            this.error = error ? error : '';
-          });
-      } else {
-        this.userService.register(this.signupForm.value)
-          .pipe(first())
-          .subscribe(
-            data => {
-              this.successmsg = true;
-              if (this.successmsg) {
-                this.router.navigate(['/account/login']);
-              }
-            },
-            error => {
-              this.error = error ? error : '';
-            });
-      }
+        });
+    }
     }
   }
-}
+
