@@ -1,20 +1,21 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 
 import { AuthenticationService } from '../../../core/services/auth.service';
 import { environment } from '../../../../environments/environment';
 import { first } from 'rxjs/operators';
 import { UserProfileService } from '../../../core/services/user.service';
-
+import { AuthenticationControllerService } from 'src/app/services/services';
+import { RegistrationRequest } from 'src/app/services/models';
 @Component({
   selector: 'app-signup',
   templateUrl: './signup.component.html',
   styleUrls: ['./signup.component.scss']
 })
 export class SignupComponent implements OnInit {
-
-  signupForm: FormGroup;
+  registerRequest:RegistrationRequest = {email: '', firstname: '', lastname: '', password: ''};
+  signupForm: UntypedFormGroup;
   submitted = false;
   error = '';
   successmsg = false;
@@ -23,12 +24,16 @@ export class SignupComponent implements OnInit {
   year: number = new Date().getFullYear();
 
   // tslint:disable-next-line: max-line-length
-  constructor(private formBuilder: FormBuilder, private route: ActivatedRoute, private router: Router, private authenticationService: AuthenticationService,
+  constructor(private formBuilder: UntypedFormBuilder,
+     private route: ActivatedRoute, 
+     private router: Router, 
+     private authService:AuthenticationControllerService,
     private userService: UserProfileService) { }
 
   ngOnInit() {
     this.signupForm = this.formBuilder.group({
       username: ['', Validators.required],
+      lastname: ['', Validators.required],
       email: ['', [Validators.required, Validators.email]],
       password: ['', Validators.required],
     });
@@ -47,30 +52,21 @@ export class SignupComponent implements OnInit {
     if (this.signupForm.invalid) {
       return;
     } else {
-      if (environment.defaultauth === 'firebase') {
-        this.authenticationService.register(this.f.email.value, this.f.password.value).then((res: any) => {
-          this.successmsg = true;
-          if (this.successmsg) {
-            this.router.navigate(['/dashboard']);
+      this.registerRequest.firstname= this.f.username.value;
+      this.registerRequest.lastname= this.f.lastname.value;
+      this.registerRequest.email= this.f.email.value;
+      this.registerRequest.password = this.f.password.value;
+      this.authService.register({
+        body: this.registerRequest
+      })
+        .subscribe({
+          next: () => {
+            this.router.navigate(['account/activiate-account']);
+          },
+          error: (err) => {
+            this.error = err.error.validationErrors;
           }
-        })
-          .catch(error => {
-            this.error = error ? error : '';
-          });
-      } else {
-        this.userService.register(this.signupForm.value)
-          .pipe(first())
-          .subscribe(
-            data => {
-              this.successmsg = true;
-              if (this.successmsg) {
-                this.router.navigate(['/account/login']);
-              }
-            },
-            error => {
-              this.error = error ? error : '';
-            });
-      }
+        });
     }
-  }
+    }
 }
