@@ -2,12 +2,11 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { LoginComponent } from './login/login.component';
-import { Login2Component } from './login2/login2.component';
-
 import { SignupComponent } from './signup/signup.component';
 import { PasswordresetComponent } from './passwordreset/passwordreset.component';
 import { Register2Component } from './register2/register2.component';
 import { Recoverpwd2Component } from './recoverpwd2/recoverpwd2.component';
+import { ActiviateAccountComponent } from './activiate-account/activiate-account.component';
 
 const routes: Routes = [
     {
@@ -31,9 +30,12 @@ const routes: Routes = [
         component: Recoverpwd2Component
     },
     {
-        path: 'login-2',
-        component: Login2Component
+        path: 'activiate-account',
+        component: ActiviateAccountComponent
     }
+
+
+   
 ];
 
 @NgModule({
