@@ -7,7 +7,12 @@ import { environment } from '../../../environments/environment';
 import { CookieService } from 'ngx-cookie-service';
 import { LanguageService } from '../../core/services/language.service';
 import { TranslateService } from '@ngx-translate/core';
+import { TokenService } from 'src/app/services/token/token.service';
 
+import {  Subscription } from 'rxjs';
+import { WebSocketService } from 'src/websocket.service';
+import { NotificationControllerService } from 'src/app/services/services';
+import { Notification } from 'src/app/services/models';
 @Component({
   selector: 'app-topbar',
   templateUrl: './topbar.component.html',
@@ -18,18 +23,29 @@ import { TranslateService } from '@ngx-translate/core';
  * Topbar component
  */
 export class TopbarComponent implements OnInit {
-
+  isUser: boolean = false;
   element;
   cookieValue;
   flagvalue;
   countryName;
   valueset;
+  firstname:string;
+  notifications:Notification[];
+  //
+  notification: { title: string, date: string ,id:Number}[] = [];
+  private subscription: Subscription;
+  messages: any[] = [];
+  unreadCount: number = 0;
+ 
 
   constructor(@Inject(DOCUMENT) private document: any, private router: Router, private authService: AuthenticationService,
               private authFackservice: AuthfakeauthenticationService,
               public languageService: LanguageService,
               public translate: TranslateService,
-              public _cookiesService: CookieService) {
+              public _cookiesService: CookieService,
+            private tokenservice:TokenService,
+            private webSocketService: WebSocketService,
+            private notifSevice: NotificationControllerService) {
   }
 
   listLang = [
@@ -46,8 +62,11 @@ export class TopbarComponent implements OnInit {
   @Output() mobileMenuButtonClicked = new EventEmitter();
 
   ngOnInit() {
+    this.isUser = this.tokenservice.hasRole('USER'); 
+    this.firstname=this.tokenservice.getFirstName().toUpperCase();
     this.openMobileMenu = false;
     this.element = document.documentElement;
+
 
     this.cookieValue = this._cookiesService.get('lang');
     const val = this.listLang.filter(x => x.lang === this.cookieValue);
@@ -57,8 +76,105 @@ export class TopbarComponent implements OnInit {
     } else {
       this.flagvalue = val.map(element => element.flag);
     }
+    //
+    this.loadUnReadNotifications();
+    this.webSocketService.connect();
+    console.log('WebSocket service connected.');
+    
+    this.subscription = this.webSocketService.getMessages().subscribe(
+      message => {
+        this.processMessage(message);
+        //this.messages.push(message);
+        this.unreadCount++;
+        console.log('New message received: ', message);
+      },
+      error => {
+        console.error('Erreur lors de la connexion au WebSocket:', error);
+      }
+    );
+  }
+  ngOnDestroy(): void {
+    // Déconnexion du WebSocket
+    this.webSocketService.disconnect();
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
   }
 
+  private processMessage(message: string): void {
+    // Regex pour extraire le titre, la date et l'id
+    const regex = /Nouvelle offre publiée : (.+?)\. Date de publication : (.+?)\. l'id: (\d+)/;
+    const match = message.match(regex);
+  
+    if (match) {
+      const title = match[1];
+      const publicationDateStr = match[2];
+      const id = match[3];
+      
+      console.log("title:", title);
+      console.log("publicationDateStr:", publicationDateStr);
+      console.log("id:", id);
+  
+      // Convertir la date en format Date
+      const publicationDate = new Date(publicationDateStr);
+      const currentDate = new Date();
+      const timeDifference = currentDate.getTime() - publicationDate.getTime();
+  
+      const formatTimeDifference = (diff: number): string => {
+        const seconds = Math.floor(diff / 1000);
+        const minutes = Math.floor(seconds / 60);
+        const hours = Math.floor(minutes / 60);
+        const days = Math.floor(hours / 24);
+  
+        if (days > 0) {
+          return `${days} jour${days > 1 ? 's' : ''} ago`;
+        } else if (hours > 0) {
+          return `${hours} heure${hours > 1 ? 's' : ''} ago`;
+        } else if (minutes > 0) {
+          return `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
+        } else {
+          return `${seconds} seconde${seconds > 1 ? 's' : ''} ago`;
+        }
+      };
+  
+      const differenceMessage = timeDifference < 60000 ? 'Now' : formatTimeDifference(timeDifference);
+  
+      // Ajouter la notification avec l'id et le titre
+      this.notification.push({ 
+        title, 
+        date: differenceMessage,
+        id:Number(id) // Ajouter l'id à la notification si nécessaire
+      });
+    }
+  }
+  
+  
+  loadUnReadNotifications(): void {
+    this.notifSevice.getUnreadNotifications().subscribe(
+      async data => {
+        if (data instanceof Blob) {
+          let jsonString = await data.text();
+          this.notifications = JSON.parse(jsonString);
+          this.notifications.forEach(notification => {
+            this.processMessage(notification.message);
+          });
+        }
+        this.unreadCount=this.notifications.length;
+      },
+      error => {
+        console.error('Erreur lors de la récupération des offres non archivées:', error);
+      }
+    );
+  }
+  markAllAsRead(): void {
+    const MAX_NOTIFICATIONS = 10;
+    this.notifSevice.markAllNotificationsAsRead().subscribe(() => {
+      const latestNotifications = this.notifications.slice(-MAX_NOTIFICATIONS);
+     
+      this.unreadCount = 0;
+      this.notifications.forEach(notification => notification.readd = true);
+    });
+  }
   setLanguage(text: string, lang: string, flag: string) {
     this.countryName = text;
     this.flagvalue = flag;
@@ -85,12 +201,14 @@ export class TopbarComponent implements OnInit {
    * Logout the user
    */
   logout() {
-    if (environment.defaultauth === 'firebase') {
-      this.authService.logout();
-    } else {
-      this.authFackservice.logout();
-    }
-    this.router.navigate(['/account/login']);
+    // if (environment.defaultauth === 'firebase') {
+    //   this.authService.logout();
+    // } else {
+    //   this.authFackservice.logout();
+    // }
+    // this.router.navigate(['/account/login']);
+    localStorage.removeItem('token');
+    window.location.reload();
   }
 
   /**
@@ -128,4 +246,9 @@ export class TopbarComponent implements OnInit {
       }
     }
   }
+  goToMyWallet() {
+    this.router.navigate(['/projects/grid']); 
+  }
+
+  
 }
