@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { UserlistComponent } from './userlist/userlist.component';
 import { UsergridComponent } from './usergrid/usergrid.component';
 import { ProfileComponent } from './profile/profile.component';
-
+import { UserlistComponent } from './userlist/userlist.component';
 const routes: Routes = [
     {
-        path: 'list',
+        path: 'list/:id',
         component: UserlistComponent
     },
     {
