@@ -7,17 +7,13 @@ import { UIModule } from '../../shared/ui/ui.module';
 import { WidgetModule } from '../../shared/widget/widget.module';
 
 import { NgApexchartsModule } from 'ng-apexcharts';
-import { SharedModule } from './saas/shared/shared.module'
 import { NgbDropdownModule, NgbTooltipModule, NgbNavModule } from '@ng-bootstrap/ng-bootstrap'
 import { SimplebarAngularModule } from 'simplebar-angular';
 
 import { DefaultComponent } from './default/default.component';
-import { SaasComponent } from './saas/saas.component';
-import { CryptoComponent } from './crypto/crypto.component';
-import { BlogComponent } from './blog/blog.component';
 
 @NgModule({
-  declarations: [DefaultComponent, SaasComponent, CryptoComponent, BlogComponent],
+  declarations: [DefaultComponent],
   imports: [
     CommonModule,
     FormsModule,
@@ -29,7 +25,6 @@ import { BlogComponent } from './blog/blog.component';
     NgbNavModule,
     WidgetModule,
     NgApexchartsModule,
-    SharedModule,
     SimplebarAngularModule
   ]
 })
