@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
+import { NgbTooltipModule ,NgbAlertModule} from '@ng-bootstrap/ng-bootstrap';
 import { InvoicesRoutingModule } from './invoices-routing.module';
 import { UIModule } from '../../shared/ui/ui.module';
-
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ListComponent } from './list/list.component';
 import { DetailComponent } from './detail/detail.component';
 
@@ -14,7 +14,9 @@ import { DetailComponent } from './detail/detail.component';
     CommonModule,
     InvoicesRoutingModule,
     UIModule,
-    NgbTooltipModule
+    NgbTooltipModule,
+    FormsModule, ReactiveFormsModule,
+    NgbAlertModule,
   ]
 })
 export class InvoicesModule { }
