@@ -6,38 +6,30 @@ import { NgbNavModule, NgbDropdownModule, NgbModalModule, NgbTooltipModule , Ngb
 import { NgApexchartsModule } from 'ng-apexcharts';
 import { FullCalendarModule } from '@fullcalendar/angular';
 import { SimplebarAngularModule } from 'simplebar-angular';
-import dayGridPlugin from '@fullcalendar/daygrid'; // a plugin
-import interactionPlugin from '@fullcalendar/interaction'; // a plugin
+import dayGridPlugin from '@fullcalendar/daygrid';
+import interactionPlugin from '@fullcalendar/interaction'; 
 import bootstrapPlugin from "@fullcalendar/bootstrap";
 import { LightboxModule } from 'ngx-lightbox';
-
+import { ProjectsModule } from './projects/projects.module';
 import { WidgetModule } from '../shared/widget/widget.module';
 import { UIModule } from '../shared/ui/ui.module';
 
 import { PagesRoutingModule } from './pages-routing.module';
 
 import { DashboardsModule } from './dashboards/dashboards.module';
-import { EcommerceModule } from './ecommerce/ecommerce.module';
-import { CryptoModule } from './crypto/crypto.module';
-import { EmailModule } from './email/email.module';
-import { InvoicesModule } from './invoices/invoices.module';
-import { ProjectsModule } from './projects/projects.module';
-import { TasksModule } from './tasks/tasks.module';
 import { ContactsModule } from './contacts/contacts.module';
-import { BlogModule } from "./blog/blog.module";
-import { UtilityModule } from './utility/utility.module';
-import { UiModule } from './ui/ui.module';
+import { InvoicesModule } from './invoices/invoices.module';
 import { FormModule } from './form/form.module';
 import { TablesModule } from './tables/tables.module';
-import { IconsModule } from './icons/icons.module';
-import { ChartModule } from './chart/chart.module';
-import { CalendarComponent } from './calendar/calendar.component';
-import { MapsModule } from './maps/maps.module';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { ChatComponent } from './chat/chat.component';
+import { DropzoneConfigInterface } from 'ngx-dropzone-wrapper';
 
-import { FilemanagerComponent } from './filemanager/filemanager.component';
-
+const DROPZONECONFIG: DropzoneConfigInterface= {
+  // URL to send the files
+  url: 'http://localhost:8080/condidature/Add',
+  maxFilesize: 50,
+  acceptedFiles: 'image/*',
+};
 FullCalendarModule.registerPlugins([ // register FullCalendar plugins
   dayGridPlugin,
   interactionPlugin,
@@ -45,7 +37,9 @@ FullCalendarModule.registerPlugins([ // register FullCalendar plugins
 ]);
 
 @NgModule({
-  declarations: [CalendarComponent, ChatComponent, FilemanagerComponent],
+  declarations: [
+   
+  ],
   imports: [
     CommonModule,
     FormsModule,
@@ -55,28 +49,17 @@ FullCalendarModule.registerPlugins([ // register FullCalendar plugins
     NgApexchartsModule,
     ReactiveFormsModule,
     DashboardsModule,
-    CryptoModule,
-    EcommerceModule,
-    EmailModule,
     InvoicesModule,
     HttpClientModule,
-    ProjectsModule,
     UIModule,
-    TasksModule,
     ContactsModule,
-    BlogModule,
-    UtilityModule,
-    UiModule,
     FormModule,
-    TablesModule,
-    IconsModule,
-    ChartModule,
     WidgetModule,
-    MapsModule,
     FullCalendarModule,
     NgbNavModule,
     NgbTooltipModule,
     NgbCollapseModule,
+    TablesModule,
     SimplebarAngularModule,
     LightboxModule
   ],
